+++ controllers/keda/scaledjob_finalizer.go
@@ -50,6 +50,9 @@ func (r *ScaledJobReconciler) finalizeScaledJob(ctx context.Context, logger logr
 			return err
 		}
 
+		if _, err := r.updateTriggerAuthenticationStatusOnDelete(ctx, logger, scaledJob); err != nil {
+			logger.Error(err, "Failed to update TriggerAuthentication Status after removing a finalizer")
+		}
 		r.updatePromMetricsOnDelete(namespacedName)
 	}