+++ controllers/keda/util/string_lists.go
@@ -16,6 +16,10 @@ limitations under the License.
 
 package util
 
+import (
+	"strings"
+)
+
 // Contains checks if the passed string is present in the given slice of strings.
 func Contains(list []string, s string) bool {
 	for _, v := range list {
@@ -34,4 +38,45 @@ func Remove(list []string, s string) []string {
 		}
 	}
 	return list
+}
+
+// AppendIntoString append a new string into a string that has seprator
+// For example,
+//
+//	-- input: `viewer,editor`, `owner`, `,`  output: `viewer,editor,owner`
+func AppendIntoString(srcStr string, appendStr string, sep string) string {
+	if appendStr == "" {
+		return srcStr
+	}
+
+	splitStrings := []string{}
+	if srcStr != "" {
+		splitStrings = strings.Split(srcStr, sep)
+	}
+
+	if !Contains(splitStrings, appendStr) {
+		splitStrings = append(splitStrings, appendStr)
+		srcStr = strings.Join(splitStrings, sep)
+	}
+	return srcStr
+}
+
+// RemoveFromString remove a string from src string that has seprator
+// For example,
+//
+//	-- input: `viewer,editor,owner`, `owner`, `,`  output: `viewer,editor`
+//	-- input: `viewer,editor,owner`, `owner`, `:`  output: `viewer,editor,owner`
+func RemoveFromString(srcStr string, str string, sep string) string {
+	if srcStr == "" {
+		return srcStr
+	}
+
+	splitStrings := []string{}
+	if srcStr != "" {
+		splitStrings = strings.Split(srcStr, sep)
+	}
+
+	splitStrings = Remove(splitStrings, str)
+	srcStr = strings.Join(splitStrings, sep)
+	return srcStr
 }