+++ pkg/scalers/authentication/authentication_types.go
@@ -6,15 +6,15 @@ import "time"
 type Type string
 
 const (
-	// APIKeyAuthType is a auth type using an API key
+	// APIKeyAuthType is an auth type using an API key
 	APIKeyAuthType Type = "apiKey"
-	// BasicAuthType is a auth type using basic auth
+	// BasicAuthType is an auth type using basic auth
 	BasicAuthType Type = "basic"
-	// TLSAuthType is a auth type using TLS
+	// TLSAuthType is an auth type using TLS
 	TLSAuthType Type = "tls"
-	// BearerAuthType is a auth type using a bearer token
+	// BearerAuthType is an auth type using a bearer token
 	BearerAuthType Type = "bearer"
-	// CustomAuthType is a auth type using a custom header
+	// CustomAuthType is an auth type using a custom header
 	CustomAuthType Type = "custom"
 	// OAuthType is a auth type using a oAuth2
 	OAuthType Type = "oauth"